requirements.txt
requests==2.28.2
loguru==0.6.0
python-telegram-bot==13.0
python-dotenv==0.21.0